arm: at91: simplify soc setup
The at91 soc support inherited a complex scheme
from the kernel.
This schme did not really give any benefits, and when
starting to migrate at91sam9263ek to use DT it started
to get in the way for a proper way to do things.

Replace the former setup with a simple function
pointer that can be called from setup.c.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Reviewed-by: Andrey Smirnov <andrew.smirnov@gmail.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>