mci: drop unused parameter from mci_switch()
The SWITCH command has two purposes:
 a) switch the command set
 b) Write to the EXT_CSD register

If the access field (bits [25:24]) in the argument are b00, we're in
case a), otherwise in b). As mci_switch() always passes
MMC_SWITCH_MODE_WRITE_BYTE (0b3) in the access field, only case b) is
relevant here. According to the eMMC specification[1] the command set
field is ignored in case b) and so the respective parameter (that is
unused already now) can be dropped.

[1] Embedded Multi-Media Card (e•MMC) Electrical Standard (5.1),
    February 2015; paragraph 6.6.1

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>