bootm: FIT: do not depend on FIT pre-opened images
When calling fit_open_configuration the FIT code already opens
the images "kernel", "ramdisk" and "dtb". This does not fit well
into the FIT code, so make the bootm code independent of these
pre-opened images so that we can drop the opening from the FIT
code in the next step.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>