ARM: return positive offset in get_runtime_offset()
When we are linked at 0x0 and running at 0x01000000 then
get_runtime_offset() should return 0x01000000 and not 0xff000000. This
makes get_runtime_offset() more consistent and better understandable.

This was tested on a Freescale i.MX53 Quickstart board. Additionally
relocate_to_adr() was tested since that is normally not called.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>