ARM: android image: Fix compiler warning on aarch64
The android image format has a u32 value which tells the loader
where to put the ATAG list. Casting this value to a pointer directly
causes a warning on aarch64. Silence it by casting it to unsigned
long first. The code is of no use on aarch64, so no need to fix anything
really.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>