ARM: imx6sx: use correct clocks for GPT
This is a stop gap solution until the upstream DT is fixed. In Linux this isn't
an issue as the driver prefers the "osc_per" clock, falling back to "per" and
then doing different things based on the SoC type and if the input clock rate
matches 3MHz. As this isn't really how the hardware works, but a fragile mix of
the clock driver exposing a clock with a faked frequency and system knowledge in
the GPT driver I prefer not to replicate this horrors in the Barebox driver,
but instead just use the system clock source.

Fixes: 2ebda864f837
       (ARM: imx: clocksource: Use per clock for determining parent freq)
Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>