commands: oftree: kill oftree -f
Being able to free the live device tree seemed like a good idea in the
early days of device tree support when the first boards were brought up.
It turned out to be illusory that noone stores pointers to the live tree
which become invalid when oftree -f is executed. Enough people stumbled
upon crashed boards with this option, so remove it finally.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>