of_node: clarify help text and do stricter commandline parsing
To create an oftree node /path/node I tried after reading the help text
of of_node:

	of_node -c /path node

similar to the syntax of of_property. This didnt result in any output
but still failed to do what I meant.

So clarify the help text to not suggest there are two parameters and
return an usage error if still two (or more) names are passed.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>