scripts/dtc: Reenable fdtget
compilation of fdtget was lost during the update to version 1.4.6.
We need this tool internally for the build process when imd support
is enabled.

Apparently our dtc code comes from the Kernel which doesn't have the
upstream version of fdtget.c and it doesn't compile. This patch
changes fdtget.c to the upstream version as of 1.4.6. This isn't
noticed in the Kernel because fdtget isn't compiled there.

Fixes: 8a8982541 scripts/dtc: Update to upstream version 1.4.6

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>