ARM: MXS: allow starting from battery input without 4P2 source enabled
Some boards don't need the 4.2V power source at all, so allow them to
keep it disabled for efficiency reasons.

For backwards compatibility, don't disable the 4P2 power source on the
existing boards which are supplied from battery input.

The POWER_USE_5V code path however always enables it up as it is needed
to supply the DC-DC converter.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>