nand command: fix null pointer when adding/deleting a bb device
Commit 40ee300 introduces info option about bad blocks. The open()
function call is therefore moved out from mark good/bad block to use the
filedescriptor in info option too. If someone tries to add/delete a bb device
the optind variable is incremented and the open() call gets a null pointer.
Fix this issue by returning when the bb device is added/deleted.

Signed-off-by: Oleg Karfich <oleg.karfich@wago.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>