ARM: i.MX: Reset OTG controller during startup
When booting from USB it's sometimes observed that the USB stack on
the remote host gets confused. This happens when we adjust the PLLs
the USB controller is hanging on. The ROM just leaves the USB controller
enabled when it passes control to the bootloader, so in case the usb
clks are enabled make sure we reset the otg  before doing any critical
clock operations.

This is currently done for i.MX50, i.MX51, i.MX53 and i.MX6, but a
similar thing can happen on other i.MXes aswell.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>