fs: tftp: hide files which are actually not present on the server
In tftp_lookup we claimed that every desired file is there. This leads
to problems when a user only tests if a file is present and makes
decisions upon this information. Rather than claiming that all files
are present do a tftp_do_open() on the files and see if it is really
there.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>