ARM: i.MX: add defconfig for the Kindle i.MX50 boards
Pin down the minimal config needed for a Kindle 4/5 device to boot into
the vendor Linux.

As the board documentation says, there is an upper limit of ~240 KB for
the bootloader partition on those devices. Using XZ image compression
instead of the default LZO reduced the file size from 214 KB to 172 KB,
so enabling that too.

Signed-off-by: Roland Hieber <rohieb@rohieb.name>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>